fix integration-test
