+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/storageengine/dataregion/memtable/AbstractMemTable.java
@@ -604,7 +604,7 @@ public abstract class AbstractMemTable implements IMemTable {
             buildChunkMetaDataForMemoryChunk(
                 measurementId,
                 timestamps[0],
-                timestamps[tvListCopy.rowCount() - 1],
+                timestamps[tvListCopy.count() - 1],
                 Collections.emptyList()));
     memChunkHandleMap
         .computeIfAbsent(measurementId, k -> new ArrayList<>())
@@ -703,7 +703,7 @@ public abstract class AbstractMemTable implements IMemTable {
               buildChunkMetaDataForMemoryChunk(
                   measurementId,
                   timestamps[0],
-                  timestamps[tvListCopy.rowCount() - 1],
+                  timestamps[tvListCopy.count() - 1],
                   Collections.emptyList()));
       memChunkHandleMap
           .computeIfAbsent(measurementId, k -> new ArrayList<>())
@@ -793,7 +793,7 @@ public abstract class AbstractMemTable implements IMemTable {
   }
 
   private long[] filterDeletedTimestamp(TVList tvList, List<TimeRange> deletionList) {
-    if (deletionList.isEmpty()) {
+    if (tvList.getBitMap() == null && deletionList.isEmpty()) {
       long[] timestamps = tvList.getTimestamps().stream().flatMapToLong(LongStream::of).toArray();
       return Arrays.copyOfRange(timestamps, 0, tvList.rowCount());
     }
@@ -805,7 +805,8 @@ public abstract class AbstractMemTable implements IMemTable {
 
     for (int i = 0; i < rowCount; i++) {
       long curTime = tvList.getTime(i);
-      if (!ModificationUtils.isPointDeleted(curTime, deletionList, deletionCursor)
+      if (!tvList.isNullValue(i)
+          && !ModificationUtils.isPointDeleted(curTime, deletionList, deletionCursor)
           && (i == rowCount - 1 || curTime != lastTime)) {
         result.add(curTime);
       }
